feat(Unassign-asset): restructure assign asset action and reducer
[#159022597]
