+++ src/_actions/allocations.actions.js
@@ -4,9 +4,7 @@ import constants from '../_constants';
 const {
   LOAD_ALLOCATIONS_SUCCESS,
   LOAD_ALLOCATIONS_FAILURE,
-  LOADING_ALLOCATIONS,
-  NEW_ALLOCATION_SUCCESS,
-  NEW_ALLOCATION_FAILURE
+  LOADING_ALLOCATIONS
 } = constants;
 
 export const loadAllocationsAction = (pageNumber, limit) => (dispatch) => {
@@ -21,14 +19,4 @@ export const loadAllocationsAction = (pageNumber, limit) => (dispatch) => {
     }));
 };
 
-export const allocateAsset = newAllocation => dispatch =>
-  axios
-    .post('allocations', newAllocation)
-    .then(response => dispatch({
-      type: NEW_ALLOCATION_SUCCESS,
-      payload: response
-    }))
-    .catch(error => dispatch({
-      type: NEW_ALLOCATION_FAILURE,
-      payload: error.message
-    }));
+export default loadAllocationsAction;
