+++ src/_reducers/asset.reducer.js
@@ -4,7 +4,9 @@ import initialState from './initialState';
 const {
   LOADING_ASSET,
   LOAD_ASSET_FAILURE,
-  LOAD_ASSET_SUCCESS
+  LOAD_ASSET_SUCCESS,
+  NEW_ALLOCATION_SUCCESS,
+  NEW_ALLOCATION_FAILURE
 } = constants;
 
 export default (state = initialState.asset, action) => {
@@ -28,6 +30,18 @@ export default (state = initialState.asset, action) => {
         errorMessage: action.payload,
         hasError: true,
         isLoading: false
+      };
+    case NEW_ALLOCATION_SUCCESS:
+      return {
+        ...state,
+        newAllocation: [...state.newAllocation, action.payload.data]
+      };
+    case NEW_ALLOCATION_FAILURE:
+      return {
+        ...state,
+        hasError: true,
+        errorMessage: action.payload,
+        newAllocation: []
       };
     default:
       return state;
